proyecto_flet/pyproject.toml
[project]
name = "proyecto-flet"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Flet developer", email = "you@example.com" }
]
dependencies = [
    "flet==0.80.5",
    "pyrebase4==4.9.0",
    "flet-geolocator==0.80.5",
    "flet-map==0.80.5",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "flet-cli==0.80.5",
    "flet-desktop==0.80.5",
    "flet-web==0.80.5",
]

[tool.flet]
org = "com.mycompany"
product = "proyecto-flet"
company = "Flet"
copyright = "Copyright (C) 2023-2026 by Flet"
permissions = ["location"]

[tool.flet.app]
path = "src"

[tool.flet.android.permission]
"android.permission.ACCESS_FINE_LOCATION" = true
"android.permission.ACCESS_COARSE_LOCATION" = true
"android.permission.ACCESS_BACKGROUND_LOCATION" = true